accel/amdxdna: split bus-agnostic core out of amdxdna_pci_drv.c
Introduce amdxdna_drv.c holding the bus-agnostic DRM driver, file ops,
client open/close and shared ioctls, leaving amdxdna_pci_drv.c with only
PCI attachment (ids, probe/remove, sriov, module_pci_driver). This lets an
auxiliary-bus-only build link the DRM core without the PCI entry.

Build wiring: single amdxdna.ko with XDNA_BUS_TYPE={pci,aux} selecting
OFT_CONFIG_AMDXDNA_PCI (amdxdna_pci_drv.o + aie*/npu*) or
OFT_CONFIG_AMDXDNA_AUX (amdxdna_aux_drv.o + ve2_*). PCI behavior unchanged.
